+++ app/src/main/java/com/amsterdam/aflami/di/viewModelModule.kt
@@ -2,7 +2,7 @@ package com.amsterdam.aflami.di
 
 import com.example.viewmodel.search.globalSearch.GlobalSearchViewModel
 import com.example.viewmodel.search.countrySearch.SearchByCountryViewModel
-import com.example.viewmodel.searchByActor.SearchByActorViewModel
+import com.example.viewmodel.search.actorSearch.SearchByActorViewModel
 import com.example.viewmodel.utils.dispatcher.DefaultDispatcherProvider
 import com.example.viewmodel.utils.dispatcher.DispatcherProvider
 import org.koin.core.module.dsl.singleOf
