+++ ui/src/main/java/com/example/ui/screens/searchByActor/SearchByActorScreen.kt
@@ -38,10 +38,10 @@ import com.example.designsystem.theme.AflamiTheme
 import com.example.designsystem.utils.ThemeAndLocalePreviews
 import com.example.ui.application.LocalNavController
 import com.example.ui.screens.searchByCountry.Loading
-import com.example.viewmodel.searchByActor.SearchByActorEffect
-import com.example.viewmodel.searchByActor.SearchByActorInteractionListener
-import com.example.viewmodel.searchByActor.SearchByActorScreenState
-import com.example.viewmodel.searchByActor.SearchByActorViewModel
+import com.example.viewmodel.search.actorSearch.SearchByActorEffect
+import com.example.viewmodel.search.actorSearch.SearchByActorInteractionListener
+import com.example.viewmodel.search.actorSearch.SearchByActorScreenState
+import com.example.viewmodel.search.actorSearch.SearchByActorViewModel
 import org.koin.androidx.compose.koinViewModel
 
 @Composable
