+++ viewModel/src/main/java/com/example/viewmodel/search/actorSearch/SearchByActorEffect.kt
@@ -1,0 +1,7 @@
+package com.example.viewmodel.search.actorSearch
+
+sealed interface SearchByActorEffect{
+
+    data object NavigateBack: SearchByActorEffect
+    data object NoInternetConnection: SearchByActorEffect
+}
