--- viewModel/src/main/java/com/example/viewmodel/searchByActor/SearchByActorEffect.kt
@@ -1,7 +1,0 @@
-package com.example.viewmodel.searchByActor
-
-sealed interface SearchByActorEffect{
-
-    data object NavigateBack:SearchByActorEffect
-    data object NoInternetConnection:SearchByActorEffect
-}
